\name{CodataGS-package}
\alias{CodataGS-package}
\alias{CodataGS}
\docType{package}
\title{
\packageTitle{CodataGS}
}
\description{
\packageDescription{CodataGS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CodataGS}
\packageIndices{CodataGS}
This package performs genomic prediction based on SNP codata. The main function is genomicEBV.w.codata.
}
\author{
\packageAuthor{CodataGS}

Maintainer: \packageMaintainer{CodataGS}
}
\keyword{ package }